Added context/reducer in product
