+++ src/App.js
@@ -1,6 +1,4 @@
-import { useState } from 'react';
 import './App.css';
-import { StoreProvider } from './Store/context';
 import Header from './Pages/Header';
 import HeroSection from './Pages/HeroSection';
 import { BrowserRouter } from 'react-router-dom';
