+++ src/Product/productReducer.js
@@ -1,9 +1,19 @@
 const ProductReducer = (state, action) => {
     switch (action.type) {
-        case "PRICE_HIGH_TO_LOW":
-            return state.sortedData.sort((a,b) => a["price"]-b["price"])
-        case "PRICE_LOW_TO_HIGH":
-            return state.sortedData.sort((a,b) => b["price"]-a["price"])
+        case "SORT": 
+            return {...state, sortBy: action.payload}
+        case "TOGGLE_INVENTORY":
+            return {...state, showInventoryAll: !state.showInventoryAll}
+        case "TOGGLE_PRIME_CHOICE":
+            return {...state, showPrimeChoice: !state.showPrimeChoice}
+        case "CATEGORY":
+            return {...state, category: action.payload}
+        case "RESET":
+            return {...state, 
+                sortBy: null,
+                showPrimeChoice: false,
+                showInventoryAll: false,
+                category: null}
         default:
             return state
     }
